카테고리 페이지의 주요 요소를 별도 컴포넌트로 분리
