+++ src/lib/organisms/Category/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./Category.svelte";
+export * from "./service";
