+++ src/routes/(main)/category/[[id]]/+page.svelte
@@ -3,10 +3,9 @@
   import { goto } from "$app/navigation";
   import { TopBar } from "$lib/components";
   import { getCategoryInfo, type CategoryInfo } from "$lib/modules/filesystem";
+  import Category from "$lib/organisms/Category";
   import { masterKeyStore } from "$lib/stores";
   import CreateCategoryModal from "./CreateCategoryModal.svelte";
-  import Files from "./Files.svelte";
-  import SubCategories from "./SubCategories.svelte";
   import { requestCategoryCreation } from "./service";
 
   let { data } = $props();
@@ -34,32 +33,16 @@
   {#if data.id !== "root"}
     <TopBar title={$info?.name} xPadding />
   {/if}
-  {#if $info}
-    <div class="flex-grow space-y-4 bg-gray-100 pb-[5.5em]">
-      <div class="space-y-4 bg-white p-4">
-        {#if data.id !== "root"}
-          <p class="text-lg font-bold text-gray-800">하위 카테고리</p>
-        {/if}
-        {#key $info}
-          <SubCategories
-            info={$info}
-            onCategoryClick={({ id }) => goto(`/category/${id}`)}
-            onCategoryCreateClick={() => {
-              isCreateCategoryModalOpen = true;
-            }}
-          />
-        {/key}
-      </div>
-      {#if data.id !== "root"}
-        <div class="space-y-4 bg-white p-4">
-          <p class="text-lg font-bold text-gray-800">파일</p>
-          {#key $info}
-            <Files info={$info} onFileClick={({ id }) => goto(`/file/${id}`)} />
-          {/key}
-        </div>
-      {/if}
-    </div>
-  {/if}
+  <div class="flex-grow bg-gray-100 pb-[5.5em]">
+    {#if $info}
+      <Category
+        info={$info}
+        onSubCategoryClick={({ id }) => goto(`/category/${id}`)}
+        onSubCategoryCreateClick={() => (isCreateCategoryModalOpen = true)}
+        onFileClick={({ id }) => goto(`/file/${id}`)}
+      />
+    {/if}
+  </div>
 </div>
 
 <CreateCategoryModal bind:isOpen={isCreateCategoryModalOpen} onCreateClick={createCategory} />
